Commit semua perubahan

+++ database/migrations/2023_10_02_093718_add_rk_id_colum_to_tugas.php
@@ -22,9 +22,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('tugas', function (Blueprint $table) {
-            //
-            $table->dropColumn('rk_id');
-        });
+        if (Schema::hasColumn('tugas', 'rk_id')) {
+            Schema::table('tugas', function (Blueprint $table) {
+                $table->dropColumn('rk_id');
+            });
+        }
     }
 };
